Add HathifileWriter class with limit on number of HTIDs in batch rights query
