+++ spec/jobs/generate_hathifile_spec.rb
@@ -107,14 +107,4 @@ RSpec.describe GenerateHathifile do
         .and match(/^job_records_processed\S*job="generate_hathifiles"\S* 1$/m)
     end
   end
-
-  describe "#batch_extract_rights" do
-    it "extracts rights_timestamp and access_profile" do
-      rights = GenerateHathifile.new.batch_extract_rights(["test.pd_google"])
-      expect(rights).to be_a(Hash)
-      expect(rights["test.pd_google"]).not_to be_nil
-      expect(rights["test.pd_google"][:rights_timestamp]).to be_a(Time)
-      expect(rights["test.pd_google"][:access_profile]).to be_a(String)
-    end
-  end
 end
